Clean up dead code and add beginner-friendly inline comments
Quick wins for code quality and documentation clarity:

- Remove unused calculate_mse function from metrics.py
- Inline calculate_ade into generate_results_summary.py (only place it's used)
- Delete metrics.py module (no longer needed)
- Fix unused ArgumentParser import in prep_data.py
- Fix README references:
  - Update notebook link to animated_results_visualization.ipynb
  - Add clear explanation of ADE metric with link to RESULTS.md
  - Document what's available in results directory
- Add educational inline comments for deep learning newcomers:
  - Explain Transformer feature choices (relative positions, velocity, etc.)
  - Document Zoo architecture grid structure with visual representation
  - Clarify data mirroring strategy for evaluation
  - Explain hyperparameter search space (model_dim, num_layers, learning_rate)
  - Provide concrete examples of tradeoffs (overfitting, capacity, etc.)

These changes eliminate confusion for newcomers while keeping the codebase
clean and focused. All linting checks pass.
